ltp-testsuite: bump to version 20160920
Patch 0001-fix-uClibc-build.patch is now part of upstream so it is not
needed in Buildroot anymore. The remaining patches were just rebased and
then renumbered so that they apply cleanly to this new version.

This has been build-tested with both glibc and uClibc toolchains and
run-tested on x86_64 and aarch64.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>