configs/freescale_p1010rdb_pa: remove U-Boot build
Binutils was locked down to 2.24 because U-Boot 2014.01 didn't build
with more recent binutils. However, this doesn't work anymore because
binutils 2.24 is deprecated.

Current U-Boot 2016.09.01 does build successfully, but we don't have a
board anymore to test it.

Since updating U-Boot is generally not needed anyway, we just remove
the U-Boot support.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Cc: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>