pppd: fix build with Linux 4.8 headers
This commit fixes the build of pppd with toolchain using Linux 4.8
headers. The patch has been taken from a pull request pending on the
Github project, which is unfortunately somewhat dead.

The fix is not the most beautiful one, but it does the job, in a piece
of source code that's anyway rather messy in terms of mixing userspace
headers and kernel headers.

Fixes:

  http://autobuild.buildroot.net/results/bdc4f274e527c5aebf9bd1b2ca85ba8e7da57e9d/
  and numerous similar other build failures

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>