qemu: fix build of target variant
Similar to what was done in commit
524b6b85595274ca13bd7d01adb61ce2dddd786c for the host variant, this
commit fixes the build of qemu's host variant. Qemu expects that the
cpp it is passed behave like cc: it passes the -c option, which is not
allowed by cpp, but is allowed for cc.

Fix that by overriding CPP when calling configure.

Fixes:

  http://autobuild.buildroot.net/results/0850e80ba62294397b44d8adea197a244a9a8205/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>