configs/kb9202_defconfig: remove defconfig
This configuration for the KwikByte KB9202 was added back in July 2008
(commit e27fef51834042a1b108e04927a2a91742a54327), and was never
updated since then. It still uses a 2.6.38 kernel, with the out of
tree patches from the former AT91 kernel maintainer. The board has
been discontinued according to its page at
http://www.kwikbyte.com/KB9202.html.

This defconfig no longer builds since the switch to gcc 5.x as the
default. Therefore, it's time to get rid of this defconfig.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Acked-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>