package/hplip: fix build by bumping to version 3.16.11
Adjust patch #2 to apply successfully.

Fixes:
http://autobuild.buildroot.net/results/bee/beea88e0848ccabcafe6b92c2a69074ed1114225/
http://autobuild.buildroot.net/results/4ba/4ba68d3e6764f395f3314c46c4d06b101c23479c/
http://autobuild.buildroot.net/results/345/3455044679a497fbee72378e59178724305720d9/
.. and many more.

The previous version 3.15.11 failed to build because of:

```
hpijs/hpcupsfax.cpp: In function 'int main(int, char**)':
prnt/hpijs/hpcupsfax.cpp:651:5: error: 'ppd_file_t' was not declared in
this scope
     ppd_file_t          *ppd;
     ^
```

`ppd_file_t` is defined in "cups/ppd.h" which was not included in 3.15.11, but
is in 3.16.11.

Signed-off-by: Jörg Krause <joerg.krause@embedded.rocks>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>