wget: bump to 1.19.1 to fix build issue
Due to the patches we have on wget 1.19, we need to
autoreconf. Unfortunately, when the autoreconfiguration process occurs
with host-gettext already built and installed, the build of wget fails
with a fairly weird error:

In file included from str-two-way.h:44:0,
                 from c-strcasestr.c:37:
./stdint.h:89:5: error: #if with no expression
 #if

As explained in http://git.net/ml/bug-gnulib-gnu/2017-01/msg00067.html
and the links pointed by this page, this is due to an incompatibility
between the newer version of gnulib used in wget, and an older .m4 file
in gettext.

In the context of Buildroot, the easiest way to avoid the issue is to
not autoreconf wget. The wget project has conveniently released a 1.19.1
release, which contains our two patches, plus just one small feature
addition. It is therefore reasonable to apply this as a solution to this
build issue.

Fixes:

  http://autobuild.buildroot.net/results/b62ac6fd5ce36453935c309e112262467cf0e3bf/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>