libuv: add dependency on BR2_TOOLCHAIN_HAS_SYNC_4
The libuv library uses __sync atomic built-ins:

$ readelf -a -W output/target/usr/lib/libuv.so.1.0.0 | grep __sync
   122: 00000000     0 NOTYPE  GLOBAL DEFAULT  UND __sync_val_compare_and_swap_4

but this is not currently taken into account in the libuv package,
causing some build failures in packages using libuv as an optional
dependency, such as janus-gateway and mosquitto.

Therefore, this commit updates the libuv package with this additional
dependency.  The reverse dependencies of libuv are also updated: luv, luvi
and moarvm.

Fixes:

 - http://autobuild.buildroot.net/results/bdaa67d763c0d8d377a04529c74f73bee7d4ccef/
   (janus-gateway)

 - http://autobuild.buildroot.net/results/2bc84ba2d1167018e2d48e5183ead22b6425dcf5/
   (mosquitto)

[Peter: drop cmake changes after cmake revert]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>