memtool: new package
memtool allows one to read and write memory mapped registers via /dev/mem.
The commands are inspired by the respective commands of the barebox
bootloader.  This is handy during driver development to inspect and modify
register settings.  It can also be used to modify regular files and
character devices (e.g. to paint to /dev/fb0).

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>
[Thomas: add entry to DEVELOPERS file.]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>