package/supertux: fix bug when compiling with PPC altivec vectorization
>From [1]:
Acording to a bug report in GCC [2]:
"You need to use -std=g++11 or undefine bool after the include of altivec.h
as context sensitive keywords is not part of the C++11 standard".

So use gnu++11 instead of c++11 only for altivec system.

This allow to revert [3] and keep Altivec vectorization.

[1] 0024465d8a7b51f7fc7c9e8581072695d869adca
[2] https://gcc.gnu.org/bugzilla/show_bug.cgi?id=58241#c3
[3] d2903aaf1314b50a8a10ce937d2b33c5c42e463e

Signed-off-by: Romain Naour <romain.naour@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>