package/kodi: re-install prompt for optical drives
In 81c438aa (package/kodi: Rework audio encoder support), the prompt for
the optical drive support was dropped, on the assumption that such
support was only needed when an audio encoder was enabled.

However, support for optical drives is also required when, surprisingly,
one wants to read a DVD (and probably a BD as well, but this was not
tested by lack of hardware).

Re-instate the prompt, so one can enable optical drive support even when
no audio encoder enabled.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Bernd Kuhls <bernd.kuhls@t-online.de>
Cc: Maxime Hadjinlian <maxime.hadjinlian@gmail.com>
Acked-by: Bernd Kuhls <bernd.kuhls@t-online.de>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>