package/qt5base: default to approved license
The only license we can act upon in Buildroot is the publicly
available license.

Qt can come under non-free licenses, but we have no access to
such licenses. The conditions to access the code under those
non-free, non-public licenses is unknown.

Besides, Qt5 (with Qt) is the only package that has this choice;
for other packages in a similar situation, we only use the free,
publicly-known licenses.

Finally, the name of the tarballs we download clearly hint that
they can only be used under the free license.

Drop the prompt to the accepted license, and make it mandatory;
packages will be adapted in follow-up patches.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Julien Corjon <corjon.j@ecagroup.com>
Cc: Peter Seiderer <ps.report@gmx.net>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>