configs/raspberrypi3: use correct CPU
We so far were using a Cortex-A7 as the CPU for the rpi3, as this was
the closest we had available, in 32-bit mode, to the real CPU used in
the rpi3.

But now we can also use the Cortex-A53 in 32-bit mode, so we use that,
which makes for an optimised build (A53 has an improved instruction set,
as compared to the A7).

Also drop the DTB overlay option, it defaults to 'y'.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Martin Bark <martin@barkynet.com>
Cc: Peter Seiderer <ps.report@gmx.net>
Tested-by: Ricardo Martincoski <ricardo.martincoski@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>