network-manager: bump version to 1.8.0
Added --disable-concheck to avoid configure failure when libcurl is not
available:

  checking for LIBCURL... no
  configure: error: --enable-concheck requires libcurl library.

One would expect the concheck functionality to be automatically disabled
if libcurl is not found. But upstream wants to be explicitly disabled if
not wanted to prevent building NetworkManager without connectivity
check support without any evident warning on systems where libcurl is
not installed. See:

  https://github.com/NetworkManager/NetworkManager/commit/faa1b85dab647baad7ce856e83f343af841aa6a5

Proper support for concheck will be added in a subsequent patch.

Signed-off-by: Vicente Olivert Riera <Vincent.Riera@imgtec.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>