x11r7: xf86-input-tslib: bump to version 1.0.0
As the update from 0.0.X to 1.0.0 suggests, this is a major update. The driver
uses X11's touch API now, instead of motion with button emulation. Furthermore
it supports multitouch when a recent version of tslib is installed.

It is not backwards compatible for very old versions of the X server. See
https://github.com/merge/xf86-input-tslib/releases for some details.

Signed-off-by: Martin Kepplinger <martink@posteo.de>
Tested-by: Bernd Kuhls <bernd.kuhls@t-online.de>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>