boost: add workaround patch for musl issue
musl has a bug in that <sched.h> defines CPU_ZERO(), which uses
memset(), but it doesn't have the prototype for it. This has been
fixed by upstream musl but until we rebuild our toolchains, let's have
a patch for Boost that works around this problem. We will of course
remove this patch once musl is updated to 1.1.17 and our toolchains
have been rebuilt.

Fixes:

  http://autobuild.buildroot.net/results/6884cff634367cb640940051a60e5e13f30f70a2/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>