package/freerdp: fix build of wayland client
Currently, FreeRDP uses pkg-config to find whether wayland is available.
When it is, it gets the CFLAGS and LDFLAGS and so on from pkg-config,
which is OK.

But I also gets the path to the wayland scanner from pkg-config. And
this points to the target one, when we really need the host one.

So we force-feed the correct path as a configure option.

Furthermore, wayland support needs libxkbcommon, so add this missing
dependency as well.

Fixes:
    http://autobuild.buildroot.org/results/bdfe233ade36a56d0f65ef3c3d1698fca548beed/

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Reviewed-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Reviewed-by: Samuel Martin <s.martin49@gmail.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>